4.21: Blog List Expansion, step 9

+++ controllers/blogs.js
@@ -30,8 +30,22 @@ blogRouter.post('/', async (request, response) => {
 });
 
 blogRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' });
+  }
+  const user = await User.findById(decodedToken.id);
+  const blog = await Blog.findById(request.params.id);
+
+  // if (user === null) {
+  //   return response.status(400).json({ error: 'user not found' });
+  // }
+
+  if (blog.user.toString() !== user?.id.toString()) {
+    return response.status(401).json({ error: 'user not authorized' });
+  }
   await Blog.findByIdAndDelete(request.params.id);
-  response.status(204).end();
+  return response.status(204).end();
 });
 
 blogRouter.put('/:id', async (request, response) => {
